+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Add `ordType` instance for `mathcomp.choice.GenTree.tree`.
 
+- Add more implicit arguments for `fsubsetP`, `fset2P` and `imfsetP`.
+
 # 0.1.0 (2018/04/26)
 
 Initial release